Fixed window of PDF print
