+++ Mollier/SAM.Core.Mollier.UI/Classes/BuiltInVisibilitySetting.cs
@@ -1,10 +1,6 @@
 ﻿using Newtonsoft.Json.Linq;
 using System;
-using System.Collections.Generic;
 using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace SAM.Core.Mollier.UI
 {
