+++ Mollier/SAM.Core.Mollier.UI/Classes/PdfDefaultSettings.cs
@@ -1,9 +1,4 @@
 ﻿using Newtonsoft.Json.Linq;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace SAM.Core.Mollier.UI
 {
