+++ Mollier/SAM.Core.Mollier.UI/Controls/BuiltInVisibilitySettingControl.cs
@@ -1,11 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
 using System.Data;
-using System.Drawing;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 
 namespace SAM.Core.Mollier.UI.Controls
