+++ Mollier/SAM.Core.Mollier.UI/Enums/ExportType.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SAM.Core.Mollier.UI
+﻿namespace SAM.Core.Mollier.UI
 {
     public enum ExportType
     {
