+++ Mollier/SAM.Core.Mollier.UI/Forms/MollierControlSettingsForm.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Windows.Forms;
 
 namespace SAM.Core.Mollier.UI
