+++ Mollier/SAM.Core.Mollier.UI/Query/NeighborhoodCount.cs
@@ -1,7 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Windows.Forms;
 
 namespace SAM.Core.Mollier.UI
 {
