+++ SAM_UI/SAM.Analytical.UI/SAM.Analytical.UI.csproj
@@ -117,7 +117,9 @@
     <Compile Include="Forms\SimulateForm.designer.cs">
       <DependentUpon>SimulateForm.cs</DependentUpon>
     </Compile>
+    <Compile Include="Modify\Insert.cs" />
     <Compile Include="Modify\AddMissingObjects.cs" />
+    <Compile Include="Query\Range.cs" />
     <Compile Include="Modify\Check.cs" />
     <Compile Include="Modify\Clean.cs" />
     <Compile Include="Modify\CreateMechanicalSystem.cs" />
